* Some overall cleanup that also eliminates all code-warnings.
* Fixed the CLI not having administrator privileges.

+++ Core/Alaris.Framework/Utility.cs
@@ -1438,7 +1438,7 @@ namespace Alaris.Framework
         /// <returns></returns>
         public static long MakeLong(int low, int high)
         {
-            return low | ((long) high << 32);
+            return (uint)low | ((long) high << 32);
         }
 
         private static readonly Random Rnd = new Random();
